Extract fetch.GitTarget
