+++ appcmd/app.go
@@ -2,6 +2,5 @@ package appcmd
 
 type App interface {
 	FullName() string
-	RemoteName() string
 	ProjectDirs() []string
 }
