+++ appcmd/fetch/fetch.go
@@ -7,21 +7,25 @@ import (
 
 	"github.com/spraints/mind-meld/appcmd"
 	"github.com/spraints/mind-meld/lmsp"
-
-	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing"
-	"github.com/go-git/go-git/v5/plumbing/object"
 )
 
+type Target interface {
+	Open() (TargetInstance, error)
+}
+
+type TargetInstance interface {
+	Add(name string, data []byte) error
+	Finish() (string, error)
+}
+
 func Run(app appcmd.App) error {
-	g, err := git.PlainOpen(".")
+	t, err := GitTarget("refs/lego/robot-inventor").Open()
 	if err != nil {
 		return err
 	}
 
 	projects, err := listProjects(app)
 
-	blobs := map[string]plumbing.Hash{}
 	for _, project := range projects {
 		data, err := readPythonProject(project)
 		if err != nil {
@@ -29,30 +33,18 @@ func Run(app appcmd.App) error {
 		}
 
 		if data != nil {
-			oid, err := createBlob(g, data)
-			if err != nil {
+			if err := t.Add(project.Name+".py", data); err != nil {
 				return fmt.Errorf("%s: %w", project.Name, err)
 			}
-			blobs[project.Name+".py"] = oid
 		}
 	}
 
-	tree, err := createTree(g, blobs)
+	msg, err := t.Finish()
 	if err != nil {
-		return err
+		return fmt.Errorf("error finishing fetch: %w", err)
 	}
 
-	commitID, err := createCommit(g, plumbing.ReferenceName(app.RemoteName()), tree,
-		"Update copy of "+app.FullName()+" python programs")
-	if err != nil {
-		return err
-	}
-
-	if commitID.IsZero() {
-		fmt.Printf("%s: no changes found.\n", app.RemoteName())
-	} else {
-		fmt.Printf("%s: created commit %v\n", app.RemoteName(), commitID)
-	}
+	fmt.Printf("%s.\n", msg)
 
 	return nil
 }
@@ -123,88 +115,4 @@ func readPythonProject(proj project) ([]byte, error) {
 		return nil, err
 	}
 	return []byte(program), nil
-}
-
-func createBlob(g *git.Repository, data []byte) (plumbing.Hash, error) {
-	obj := g.Storer.NewEncodedObject()
-	obj.SetType(plumbing.BlobObject)
-
-	w, err := obj.Writer()
-	if err != nil {
-		return plumbing.ZeroHash, err
-	}
-
-	n, err := w.Write(data)
-	if err != nil {
-		return plumbing.ZeroHash, err
-	}
-	if n != len(data) {
-		return plumbing.ZeroHash, fmt.Errorf("incomplete write")
-	}
-
-	if err := w.Close(); err != nil {
-		return plumbing.ZeroHash, err
-	}
-
-	return g.Storer.SetEncodedObject(obj)
-}
-
-type treeID = plumbing.Hash
-
-func createTree(g *git.Repository, blobs map[string]plumbing.Hash) (treeID, error) {
-	var tb treeBuilder
-	for name, oid := range blobs {
-		tb.Add(name, oid)
-	}
-	return tb.Build(g)
-}
-
-func createCommit(g *git.Repository, refName plumbing.ReferenceName, tree treeID, commitMsg string) (plumbing.Hash, error) {
-	// Check the ref.
-	// If the tree is the same, there's nothing to do.
-	// If the ref is there, use its OID as the parent commit.
-	var parentHashes []plumbing.Hash
-	if ref, err := g.Reference(refName, false); err == nil {
-		c, err := g.CommitObject(ref.Hash())
-		if err != nil {
-			return plumbing.ZeroHash, err
-		}
-		if c.TreeHash == tree {
-			return plumbing.ZeroHash, nil
-		}
-		parentHashes = append(parentHashes, ref.Hash())
-	}
-
-	// Get an author and committer to use for the commit.
-	var o git.CommitOptions
-	if err := o.Validate(g); err != nil {
-		return plumbing.ZeroHash, err
-	}
-
-	// Build the commit.
-	var c object.Commit
-	c.Author = *o.Author
-	c.Committer = *o.Committer
-	c.Message = commitMsg
-	c.TreeHash = tree
-	c.ParentHashes = parentHashes
-
-	// Save the commit.
-	obj := g.Storer.NewEncodedObject()
-	if err := c.Encode(obj); err != nil {
-		return plumbing.ZeroHash, err
-	}
-
-	commitID, err := g.Storer.SetEncodedObject(obj)
-	if err != nil {
-		return plumbing.ZeroHash, err
-	}
-
-	if os.Getenv("FOO") == "FOO" {
-		w, _ := g.Worktree()
-		w.Commit("", nil)
-	}
-
-	// Update the reference.
-	return commitID, g.Storer.SetReference(plumbing.NewHashReference(refName, commitID))
 }
