+++ apps/mindstormsapp/app.go
@@ -16,10 +16,6 @@ func (*App) FullName() string {
 	return "LEGO MINDSTORMS Inventor"
 }
 
-func (*App) RemoteName() string {
-	return "refs/lego/robot-inventor"
-}
-
 func (*App) ProjectDirs() []string {
 	home := os.Getenv("HOME")
 	return []string{
